requirements.txt
numpy
Pillow
PyYAML
tqdm
matplotlib
torch
torchvision
